Revert "move DashboardLayout to the Layout folder"
This reverts commit 827d8be1949604ae1475d25ab20a2d5e4571ae69.
