+++ src/containers/dashboard/DashboardLayout.tsx
@@ -3,6 +3,7 @@ import {
   SchoolOutlined,
   StarOutlined
 } from '@mui/icons-material'
+import { Typography } from '@mui/material'
 import { ReactNode } from 'react'
 
 import {
